==> Jupyter-Notebooks/3_Data_Profiling.ipynb <==
pip install ydata_profiling

# from google.colab import drive
# drive.mount('/content/drive')

import pandas as pd
from ydata_profiling import ProfileReport

# Load raw datasets
merged = pd.read_csv('/content/2_merged_uiuc_courses_gpa.csv', low_memory=False)

# Profile the merged data
merged_profile = ProfileReport(
    merged,
    title="Merged UIUC Data Profile",
    explorative=True
)
merged_profile.to_file('uiuc_merged_profile.html')

print("Merged profile → /mnt/data/uiuc_merged_profile.html")

import pandas as pd
from ydata_profiling import ProfileReport

# Load raw datasets
merged = pd.read_csv('/content/2_merged_uiuc_courses_gpa.csv', low_memory=False)

# Profile the merged data
merged_profile = ProfileReport(
    merged,
    title="Merged UIUC Data Profile",
    explorative=True
)
merged_profile.to_file('uiuc_merged_profile.html')

print("Merged profile → /mnt/data/uiuc_merged_unclean_profile.html")

# Load raw datasets
cleaned = pd.read_csv('/content/merged_cleaned_final.csv', low_memory=False)

# Profile the merged data
cleaned_profile = ProfileReport(
    merged,
    title="Cleaned UIUC Data Profile",
    explorative=True
)
cleaned_profile.to_file('uiuc_cleaned_profile.html')

print("Merged profile → /mnt/data/uiuc_merged_cleaned_profile.html")
